route: move pkg/datapath/route to pkg/datapath/linux/route
[ upstream commit 443b12724b9facd92f2399862b9fe09259547381 ]

This prepares the route utility layer for the upcoming pkg/datapath/linux
package. It's primarily a code move and renames the main update and delete
functio to be easier to undertand:

route.ReplaceRoute() -> route.Upsert()
Route.DeleteRoute() -> route.Delete()
