+++ pkg/datapath/linux/route/route.go
@@ -11,16 +11,58 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-//
-// +build darwin
 
 package route
 
 import (
 	"fmt"
+	"net"
 
-	"github.com/cilium/cilium/pkg/mtu"
+	"github.com/cilium/cilium/pkg/logging/logfields"
+	"github.com/sirupsen/logrus"
+
+	"github.com/vishvananda/netlink"
 )
+
+type Route struct {
+	Prefix  net.IPNet
+	Nexthop *net.IP
+	Local   net.IP
+	Device  string
+	MTU     int
+	Scope   netlink.Scope
+}
+
+// LogFields returns the route attributes as logrus.Fields map
+func (r *Route) LogFields() logrus.Fields {
+	return logrus.Fields{
+		"prefix":            r.Prefix,
+		"nexthop":           r.Nexthop,
+		"local":             r.Local,
+		logfields.Interface: r.Device,
+	}
+}
+
+func (r *Route) getLogger() *logrus.Entry {
+	return log.WithFields(r.LogFields())
+}
+
+// ByMask is used to sort an array of routes by mask, narrow first.
+type ByMask []Route
+
+func (a ByMask) Len() int {
+	return len(a)
+}
+
+func (a ByMask) Less(i, j int) bool {
+	lenA, _ := a[i].Prefix.Mask.Size()
+	lenB, _ := a[j].Prefix.Mask.Size()
+	return lenA > lenB
+}
+
+func (a ByMask) Swap(i, j int) {
+	a[i], a[j] = a[j], a[i]
+}
 
 // ToIPCommand converts the route into a full "ip route ..." command
 func (r *Route) ToIPCommand(dev string) []string {
@@ -37,15 +79,4 @@ func (r *Route) ToIPCommand(dev string) []string {
 	}
 	res = append(res, "dev", dev)
 	return res
-}
-
-// ReplaceRoute adds or replaces the specified route if necessary. Does nothing
-// for Darwin-based builds.
-func ReplaceRoute(route Route, mtuConfig mtu.Configuration) error {
-	return nil
-}
-
-// DeleteRoute removes a route. Does nothing for Darwin-based builds.
-func DeleteRoute(route Route) error {
-	return nil
 }
