+++ pkg/datapath/linux/route/route_darwin.go
@@ -1,0 +1,34 @@
+// Copyright 2016-2018 Authors of Cilium
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// +build darwin
+
+package route
+
+import (
+	"fmt"
+
+	"github.com/cilium/cilium/pkg/mtu"
+)
+
+// Replace adds or replaces the specified route if necessary. Does nothing
+// for Darwin-based builds.
+func Replace(route Route, mtuConfig mtu.Configuration) error {
+	return fmt.Errorf("Operation not supported ")
+}
+
+// Delete removes a route. Does nothing for Darwin-based builds.
+func Delete(route Route) error {
+	return fmt.Errorf("Operation not supported ")
+}
